requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.6
markupsafe==3.0.2
packaging==25.0
pillow==11.3.0
pyjwt==2.10.1
pytz==2025.2
pyyaml==6.0.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
